Basic annotations - floor

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,7 +5,7 @@ This module contains a type-annotated function named floor
 import math
 
 
-def floor(n: int) -> int:
+def floor(n: float) -> int:
     """
     function floor takes a float n as argument and
         returns the floor of the float.
